requirements.txt
numpy>=1.11.1
scipy>=0.18.1
pydot2>=1.0.33
ipython>=5.10.3
pyqt>=5.6.0
matplotlib>=1.5.3
